fix(gui): fix blockly style bugs (#50)
AmpMod's blocks UI would look weird compared to Scratch (and LibreKitten, which this fix was ported from), such as report bubbles being collapsed, and context menus being scrollable. Because of modern-normalize, we have to explictly override its opinionated `box-sizing: border-box;`, which I think causes this bug.

As I said earlier, this fix was ported from LibreKitten (which has a similar setup to AmpMod). It should be fine to port it over, as it was all original code by me. You don't have to credit LibreKitten, but you can if you want to.

Reviewed-on: https://codeberg.org/ampmod/ampmod/pulls/50
Co-authored-by: ACat <acat@noreply.codeberg.org>
Co-committed-by: ACat <acat@noreply.codeberg.org>